\name{.compute.ev}
\alias{.compute.ev}
\title{Simulate Expected Value for Multinomial Logit...}
\usage{.compute.ev(fitted, constraints, all.coef, x, ndim, cnames)}
\description{Simulate Expected Value for Multinomial Logit}
\value{a matrix of simulated values}
\author{Matt Owen and Olivia Lau and Kosuke Imai \email{mowen@iq.harvard.edu}}
\arguments{\item{fitted}{a fitted model object}
\item{constraints}{a constraints object}
\item{all.coef}{all the coeficients}
\item{x}{a setx object}
\item{ndim}{an integer specifying the number of dimensions}
\item{cnames}{a character-vector specifying the names of the columns}}
