\name{.compute.pv}
\alias{.compute.pv}
\title{Simulate Predicted Values...}
\usage{.compute.pv(fitted, ev, ynames)}
\description{Simulate Predicted Values}
\value{a vector of simulated values}
\author{Matt Owen and Olivia Lau and Kosuke Imai \email{mowen@iq.harvard.edu}}
\arguments{\item{fitted}{a fitted model object}
\item{ynames}{???}}
