\name{.construct.v}
\alias{.construct.v}
\title{Split Names of Vectors into N-vectors...}
\usage{.construct.v(constraints, ndim)}
\description{Split Names of Vectors into N-vectors
This function is used to organize how variables are spread
across the list of formulas}
\value{a list of character-vectors}
\author{Matt Owen and Olivia Lau and Kosuke Imai}
\arguments{\item{constraints}{a constraints object}
\item{ndim}{}}
