\name{.prepare.frame}
\alias{.prepare.frame}
\title{Correctly Prepare a data...}
\usage{.prepare.frame(obj, x)}
\description{Correctly Prepare a data.frame to Be Used with predict.mnp
param obj a zelig object
param x a setx object
value a single-row data.frame with NA's for predicted variables
and the correct values for explanatory variables (set via
the setx object)}
